+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.11.0
+
+- Updated to pnpm version 10
+- Updated node to v22
+- Updated all dependencies to the latest major version.
+- Updated GitHub Actions
+- Remove react-csv and replace with own component for CSV export
+
 ## 0.10.0
 
 - Updated to pnpm version 9